rund: add exec command
The exec command allows pid 1 to execute another program.
The necessary usecase for this is that when an open file of rund
is removed/upgraded, then rund holds the filesystems busy for writing
until it exits. Due to this, the root filesystem may not proper unmount
in case rund itself or a shared library it uses is replaced.