rund runc: do not translate username in pid 1
user name translation may involve complicated things, like LDAP queries.
Doing this in pid 1 creates a danger of pulling in a lot of dependencies.

The translation is offloaded to runc (or any other client) now.
The name & group lookup to be performed when spawning services
is left intact, as it does not happen in pid 1!