rund: add 'reload' command
This commit introduces the 'reload' command, that does immediately
respawn a postponed or periodical service.
This is like sending SIGHUP.
'reload' supersedes SIGHUP in that 'reload' is available to
non-root users also (it is then limited to the user's own services)
and it can address a limited set of services.

Signed-off-by: Kurt Van Dijck <dev.kurt@vandijck-laurijssen.be>