rund: fix oneshot mode
This commit will improve the oneshot mode services.
A oneshot mode service will only be duplicated using 'add'
if the service is already executing, by ignoring the FL_ONESHOT
and setting FL_REMOVE upon start of the service.
This truely makes is execute once, and again if a new one is scheduled
while the previous equal one still executes.

The problem I had is that when different sources schedule an equal
FL_ONESHOT service with DELAYed execution, each one yields a new
service, which is not the expected behaviour.